Extend playlist selection improvements when play main title option is selected in settings.
This moves most of the simple menu code (other than the pure GUI part) out of CGUISimpleDialog into CDiscDirectoryHelper.

It also means that the main title is selected in CDiscDirectoryHelper/CBlurayDirectory and not in DVDInputStreamBluray - this allows episodes to benefit from the enhanced selection logic rather than just playing the longest title.